+++ dsmr_datalogger/scripts/dsmr_datalogger_api_client.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 """
     https://dsmr-reader.readthedocs.io/en/latest/installation/datalogger.html
 
@@ -124,9 +125,9 @@ def main():  # noqa: C901
     if DATALOGGER_INPUT_METHOD == 'serial':
         serial_kwargs.update(dict(
             url_or_port=decouple.config('DATALOGGER_SERIAL_PORT'),
-            baudrate=decouple.config('DATALOGGER_SERIAL_BAUDRATE', cast=int),
-            bytesize=serial.EIGHTBITS,
-            parity=serial.PARITY_NONE,
+            baudrate=decouple.config('DATALOGGER_SERIAL_BAUDRATE', cast=int, default=115200),
+            bytesize=decouple.config('DATALOGGER_SERIAL_BYTESIZE', cast=int, default=serial.EIGHTBITS),
+            parity=decouple.config('DATALOGGER_SERIAL_PARITY', cast=str, default=serial.PARITY_NONE),
             stopbits=serial.STOPBITS_ONE,
             xonxoff=1,
             rtscts=0,